// kodikosERP-master/src/app/models/formatki/warehouseform.model.ts
export interface Warehouseform {
  comments?: boolean;
  status?: boolean;
  buttonDodajKontr?: boolean;
  buttonOfertaKontr?: boolean;
  buttonEdytujKontr?: boolean;
  buttonUsunKontr?: boolean;
  projekty?: boolean;
  firmy?: boolean;
  widziWszystkie?: boolean;
  pliki?: boolean;
  buttonPliki?: boolean;
}